+++ libs/database/tables/teams/team.c
