+++ libs/database/tables/users_x_teams/users_x_teams.h
